Added logging to fail handlers
I was having troubles getting my contracts published on CI, after much hair being pulled out I resorted to adding logs to publisher.js. Turned out it was failing because I made a typo in the provider name and the broker was returning the provider has not been created error.

To prevent other people from struggling with this I've added logs to the fail branches of the promises.